requirements.txt
flask
yfinance
python-dotenv
ccxt
requests
